updates on the exercise scripts
